requirements.txt
beautifulsoup4==4.12.3
greenlet==3.0.3
html5lib==1.1
iniconfig==2.0.0
lxml==5.2.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
webencodings==0.5.1
